+++ R/stepwise.R
@@ -280,6 +280,8 @@ stepwise <- function(formula,
   multico_x <- getMulticolX(data_train, x_name_orig, tolerance)
   merged_multico_x <- getMergedVar(multico_x)
   x_name <- setdiff(x_name_orig, multico_x)
+
+  dummy_name_list <- dummyNameList(x_name, data_train)
   
   result <- list()
   ## table1
@@ -291,7 +293,7 @@ stepwise <- function(formula,
   result$variable <- table2_class_table
   
   ## table3
-  table3_process <- getTable3ProcessSummary(data_train=data_train, data_test=data_test, type, strategy, metric, sle, sls, weight, x_name, y_name, intercept, include, best_n, test_method, sigma_value, num_digits, feature_ratio)
+  table3_process <- getTable3ProcessSummary(data_train=data_train, data_test=data_test, type, strategy, metric, sle, sls, weight, x_name, y_name, intercept, include, best_n, test_method, sigma_value, num_digits, feature_ratio, dummy_name_list)
   x_final_model_metric <- table3_process$final_variable
   result <- append(result,table3_process[which(names(table3_process) != "final_variable")])